[nemo-qml-plugin-systemsettings] Cope better with udisks2 invalidating properties. JB#62852
At least with new udisks2 there are property change signals with
empty updated values map and non-empty invalidated properties list
with org.freedesktop.DBus.Properties.PropertiesChanged.

The handler code was assuming an empty updated values map meant that the
instance is getting removed. It was also assuming that the update
contained all the properties, not only the changed ones.